tools/checkmdi/requirements.txt
requests
pandas
dnspython